+++ KubeArmor/common/bpffs.go
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Authors of KubeArmor
 
+//go:build linux
+
 package common
 
 // This implementation for automounting bpffs has been inspired from
