+++ KubeArmor/utils/bpflsmprobe/probe.go
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Authors of KubeArmor
 
+//go:build linux
+
 // Package probe checks whether the probed LSM support is available.
 package probe
